+++ src/main/java/com/easypark/solutionsback/controller/VacancyController.java
@@ -1,17 +1,22 @@
 package com.easypark.solutionsback.controller;
 
+import com.easypark.solutionsback.dto.request.PositionVacancyRequestDTO;
 import com.easypark.solutionsback.dto.request.VacancyRequestDTO;
 import com.easypark.solutionsback.dto.response.StatusVacancyResponseDTO;
 import com.easypark.solutionsback.service.VacancyService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.support.DefaultMessageSourceResolvable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @RestController
-@RequestMapping("/api-vacancy")
+@RequestMapping("/vacancy")
 @RequiredArgsConstructor
-public class WebVacancyController {
+public class VacancyController {
 
     private final VacancyService vacancyService;
 
@@ -21,12 +26,35 @@ public class WebVacancyController {
     }
 
     @PostMapping("/create-vacancy")
-    public @ResponseBody ResponseEntity<String> createVacancy(@RequestBody VacancyRequestDTO body) {
+    public @ResponseBody ResponseEntity<String> createVacancy(@Valid @RequestBody PositionVacancyRequestDTO body, BindingResult result) {
+        if (result.hasErrors()) {
+            String errors = result.getAllErrors().stream()
+                    .map(DefaultMessageSourceResolvable::getDefaultMessage)
+                    .collect(Collectors.joining(", "));
+            return ResponseEntity.badRequest().body(errors);
+        }
         return this.vacancyService.createVacancy(body);
     }
 
-    @PostMapping("/delete-vacancy")
-    public @ResponseBody ResponseEntity<String> deleteVacancy(@RequestBody VacancyRequestDTO body) {
+    @DeleteMapping("/delete-vacancy")
+    public @ResponseBody ResponseEntity<String> deleteVacancy(@Valid @RequestBody PositionVacancyRequestDTO body, BindingResult result) {
+        if (result.hasErrors()) {
+            String errors = result.getAllErrors().stream()
+                    .map(DefaultMessageSourceResolvable::getDefaultMessage)
+                    .collect(Collectors.joining(", "));
+            return ResponseEntity.badRequest().body(errors);
+        }
         return this.vacancyService.deleteVacancy(body);
+    }
+
+    @PutMapping("/update-status-vacancy")
+    public @ResponseBody ResponseEntity<String> updateStatusVacancy(@Valid @RequestBody List<VacancyRequestDTO> body, BindingResult result) {
+        if (result.hasErrors()) {
+            String errors = result.getAllErrors().stream()
+                    .map(DefaultMessageSourceResolvable::getDefaultMessage)
+                    .collect(Collectors.joining(", "));
+            return ResponseEntity.badRequest().body(errors);
+        }
+        return this.vacancyService.updateStatusVacancy(body);
     }
 }
